backend/requirements.txt
boto3==1.35.0
botocore==1.35.0
pydantic==2.9.0
python-jose[cryptography]==3.3.0
python-dateutil==2.9.0
pytest==8.3.0
pytest-mock==3.14.0
requests==2.32.3
numpy==1.24.3
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
langchain-text-splitters==0.2.0
tiktoken==0.8.0
twilio==8.10.0

